requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.21
celery==3.1.19
click==6.6
Django==1.8.3
django-celery==3.1.17
django-cors-headers==1.1.0
django-filter==0.13.0
django-model-utils==2.3.1
djangorestframework==3.2.3
djangorestframework-jwt==1.8.0
Jinja2==2.8
kombu==3.0.35
livereload==2.4.1
Markdown==2.6.6
MarkupSafe==0.23
mkdocs==0.15.3
mkdocs-bootstrap==0.1.1
mkdocs-bootswatch==0.4.0
mysqlclient==1.3.6
Pillow==2.9.0
psycopg2==2.6.1
PyJWT==1.4.2
pypng==0.0.18
PyQRCode==1.1
pytz==2015.7
PyYAML==3.11
requests==2.11.0
six==1.9.0
tornado==4.4.1
